Add contentful client with support for local mock data
It's useful to have some static data for the 'Events' content type to play around with. This PR aims to provide a recorded API response from a static JSON file, via an API wrapper around the `contentful` client.

The wrapper for the [`contentful` client](https://github.com/contentful/contentful.js) does a couple of things:
- provides an `allEvents()` method that returns all events
- returns only the `.fields` values as presumably we only care about those (see 'data.json')
- provides a mock client which returns static data for the 'Event' content type from a JSON file ('data.json', derived from a real response for a toy 'Event' content type).

To make a real request against the API, the env vars `SPACE` and `ACCESS_TOKEN` will need to be set, and the `ContentfulClient` instance will have to be created with `fetchStaticData=false`.