Use a config file for contentful-export
During the export we skip most types so it’s more succinct to add this to a config file.
Also rename the destination file to content-types.json, we always want to write to the same file to track the differences in git.

The change in versions in content-types.json is from manually editing fields and running migrations (and then undoing them)